Adicionado Insert Rota

+++ src/02-Domain/Interface/IRepositories/IRotaRepository.cs
@@ -3,7 +3,7 @@ using RotaDeViagem.Shared.Interfaces.IRepositories;
 
 namespace RotaDeViagem.Domain.Interface.IRepositories
 {
-    public interface IRotaRepository: IRepository<Rota, long>
+    public interface IRotaRepository: IRepository<Rota, int>
     {
     }
 }
