+++ src/04-Shared/Interfaces/IEntities/IAuditableEntity.cs
@@ -8,7 +8,7 @@ namespace RotaDeViagem.Shared.Interfaces.IEntities
 {
     public interface IAuditableEntity : IEntity
     {
-        long Id { get; set; }
+        int Id { get; set; }
         Guid UniqueId { get; set; }
         string CreatedBy { get; set; }
         string ModifiedBy { get; set; }
